Refactor: Encapsulate model preparation and loading into a dedicated class
Moved the preparation of the model and loading logic into a specific class to encapsulate functionality and reduce the number of imports.
